% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campyl407}
\alias{campyl407}
\title{A dataset containing the number of campylbactoria virus cases.}
\format{
KCWC station: A data frame with 1248 rows and 1 variables:
\describe{
\item{dta2.ts}{number of hot hours per day}
}
}
\source{
Obtained from Dewey https://docs.deweydata.io/docs/custom-weather-weather-data.
}
\usage{
campyl407
}
\description{
A dataset containing the number of campylbactoria virus cases.
}
\keyword{datasets}
