% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctsc-package.R
\docType{package}
\name{gctsc-package}
\alias{gctsc-package}
\title{gctsc: Copula Count Time Series Models}
\description{
Provides likelihood-based inference for Gaussian and Student-t copula models for univariate count time series. Supports Poisson, negative binomial, binomial, beta-binomial, and zero-inflated marginals with ARMA dependence structures. Includes simulation, maximum-likelihood estimation, residual diagnostics, and predictive inference. Implements Time Series Minimax Exponential Tilting (TMET) \doi{10.1016/j.csda.2026.108344}, an adaptation of minimax exponential tilting of Botev (2017) \doi{10.1111/rssb.12162}. Also provides a linear-cost implementation of the Geweke–Hajivassiliou–Keane (GHK) simulator following Masarotto and Varin (2012) \doi{10.1214/12-EJS721}, and the Continuous Extension (CE) approximation of Nguyen and De Oliveira (2025) \doi{10.1080/02664763.2025.2498502}. The package follows the S3 design philosophy of 'gcmr' but is developed independently.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/QNNHU/gctsc}
  \item Report bugs at \url{https://github.com/QNNHU/gctsc/issues}
}

}
\author{
\strong{Maintainer}: Quynh Nguyen \email{nqnhu2209@gmail.com}

Authors:
\itemize{
  \item Victor De Oliveira
}

}
