% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print.R
\name{print.gctsc}
\alias{print.gctsc}
\title{Print a gctsc Model Object}
\usage{
\method{print}{gctsc}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class \code{gctsc}.}

\item{digits}{Number of significant digits to display.}

\item{...}{Ignored. Included for S3 method compatibility.}
}
\value{
Returns the object \code{x} invisibly.
Called for its side effect of printing a formatted summary of the model fit.
}
\description{
Displays the call, estimation method, parameter estimates, and likelihood information.
}
