% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print.R
\name{print.summary.gctsc}
\alias{print.summary.gctsc}
\title{Print Summary of a gctsc Model}
\usage{
\method{print}{summary.gctsc}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{summary.gctsc}.}

\item{digits}{Number of significant digits to display.}

\item{...}{Ignored. Included for S3 method compatibility.}
}
\description{
Displays summary statistics and model fit information for a fitted \code{gctsc} model.
}
