% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campyl}
\alias{campyl}
\title{Weekly Campylobacter case counts across Germany}
\format{
A data frame with 1248 rows and 413 variables.
}
\usage{
data("campyl")
}
\description{
Weekly Campylobacter case counts across Germany
}
\keyword{datasets}
