% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KCWC}
\alias{KCWC}
\title{Daily aggregated weather measurements for KCWC station}
\format{
A data frame with 2665 rows and 4 variables.
}
\usage{
data("KCWC")
}
\description{
Daily aggregated weather measurements for KCWC station
}
\keyword{datasets}
